fix(50365445): slate editor fix
