+++ packages/modules.editor/src/ui/Editor.tsx
@@ -1,34 +1,19 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 import React, { useCallback, useMemo, useState } from 'react';
-import { createPortal } from 'react-dom';
 
 import { Transforms, Editor as SlateEditor, Descendant } from 'slate';
-import { Slate, Editable, ReactEditor, RenderElementProps } from 'slate-react';
+import { Slate, ReactEditor } from 'slate-react';
 
-import {
-  DndContext,
-  DragEndEvent,
-  DragOverlay,
-  PointerSensor,
-  useSensor,
-  useSensors,
-} from '@dnd-kit/core';
-import { SortableContext, verticalListSortingStrategy } from '@dnd-kit/sortable';
-import { restrictToVerticalAxis } from '@dnd-kit/modifiers';
+import { DragEndEvent, PointerSensor, useSensor, useSensors } from '@dnd-kit/core';
 
 // Заменяем импорт из @xipkg/slatetypes на локальные типы
 import { CustomElement } from '../types';
 
 // Импорты из локальных модулей
+import { EditorContent } from './EditorContent';
 import { isImageUrl } from '../utils/isUrl';
-import { RenderElement } from '../ui/elements/RenderElement';
 import { createNode } from '../utils/createNode';
-import { SortableElement } from '../ui/components/SortableElement';
-import { InlineToolbar } from '../ui/components/InlineToolbar';
-import { Leaf } from '../ui/components/Leaf';
-import { useDecorateCode } from '../hooks/useDecorateCode';
 import { codeEditorInsertText } from '../utils/codeEditorInsertText';
-import DragOverlayContent from '../ui/components/DragOverlayContent';
 import { useCollaborativeEditing } from '../hooks/useCollaborativeEditing';
 
 // Удаляем дублирующее определение CustomElement, так как оно уже есть в типах
@@ -43,100 +28,6 @@ type EditorPropsT = {
   readOnly?: boolean;
   documentName?: string;
   serverUrl?: string;
-};
-
-// Новый компонент-обёртка, который будет использовать хуки внутри контекста Slate
-const EditorContent = ({
-  editor,
-  isEditorReadOnly,
-  handleOnKeyDown,
-  items,
-  sensors,
-  clearSelection,
-  handleOnDragEnd,
-  draggingElementId,
-  setDraggingElementId,
-  activeElement,
-}: {
-  editor: ReactEditor;
-  isEditorReadOnly: boolean;
-  handleOnKeyDown?: (event: React.KeyboardEvent<HTMLDivElement>) => void;
-  items: string[];
-  sensors: ReturnType<typeof useSensors>;
-  clearSelection: () => void;
-  handleOnDragEnd: (event: DragEndEvent) => void;
-  draggingElementId: string;
-  setDraggingElementId: (id: string) => void;
-  activeElement?: CustomElement;
-}) => {
-  // Теперь эти хуки используются внутри контекста Slate
-  const decorateCode = useDecorateCode();
-
-  // Используем draggingElementId для определения состояния перетаскивания
-  const isDragging = useMemo(() => Boolean(draggingElementId), [draggingElementId]);
-
-  const renderElement = useCallback(
-    (props: RenderElementProps) => {
-      const isTopLevel = ReactEditor.findPath(editor, props.element).length === 1;
-
-      return isTopLevel ? (
-        <SortableElement {...props} renderElement={RenderElement} />
-      ) : (
-        <RenderElement {...props} />
-      );
-    },
-    [editor],
-  );
-
-  if (isEditorReadOnly) {
-    return (
-      <Editable
-        readOnly
-        className="flex flex-col gap-2 p-2 text-gray-100 focus-visible:outline-none focus-visible:[&_*]:outline-none"
-        renderElement={renderElement}
-        // @ts-ignore
-        decorate={decorateCode}
-        renderLeaf={(props) => <Leaf {...props} />}
-      />
-    );
-  }
-
-  return (
-    <DndContext
-      onDragStart={(event) => {
-        if (event.active) {
-          clearSelection();
-          setDraggingElementId(`${event.active.id}`);
-        }
-      }}
-      onDragEnd={handleOnDragEnd}
-      onDragCancel={() => {
-        setDraggingElementId('');
-      }}
-      modifiers={[restrictToVerticalAxis]}
-      sensors={sensors}
-    >
-      <SortableContext items={items} strategy={verticalListSortingStrategy}>
-        <InlineToolbar />
-        <Editable
-          onKeyDown={handleOnKeyDown}
-          className={`flex flex-col gap-2 p-2 text-gray-100 focus-visible:outline-none focus-visible:[&_*]:outline-none ${
-            isDragging ? 'cursor-grabbing' : ''
-          }`}
-          renderElement={renderElement}
-          renderLeaf={(props) => <Leaf {...props} />}
-          // @ts-ignore
-          decorate={decorateCode}
-        />
-      </SortableContext>
-      {createPortal(
-        <DragOverlay>
-          {activeElement && <DragOverlayContent element={activeElement} />}
-        </DragOverlay>,
-        document.body,
-      )}
-    </DndContext>
-  );
 };
 
 export const Editor = ({
@@ -194,10 +85,8 @@ export const Editor = ({
         .then((text) => {
           if (text) {
             const { selection } = editor;
-
             if (selection) {
               const [parentNode] = SlateEditor.parent(editor, selection);
-
               if (parentNode && (parentNode as CustomElement)?.type === 'code') {
                 Transforms.insertText(editor, text);
               }
